Restructure Function component to match design.
