+++ Ordersystem/Ordersystem/Ordersystem.Droid/Ordersystem.Droid.csproj
@@ -61,7 +61,7 @@
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utilities\SQLite_Android.cs" />
-    <Compile Include="Utilities\McsSqlConnection.cs" />
+    <Compile Include="Utilities\MCSManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
