+++ Ordersystem/Ordersystem/Ordersystem/Ordersystem.csproj
@@ -34,7 +34,6 @@
   <ItemGroup>
     <Compile Include="Enums\Enumerations.cs" />
     <Compile Include="Exceptions\ItemNotFoundException.cs" />
-    <Compile Include="Functions\CommunicationManager.cs" />
     <Compile Include="Model\Customer.cs" />
     <Compile Include="Model\DayMenu.cs" />
     <Compile Include="Model\DayMenuSelection.cs" />
@@ -48,6 +47,7 @@
     <Compile Include="Model\Orderlist.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <None Include="packages.config" />
+    <Compile Include="Functions\LocalManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="MySql.Data">
